requirements.txt
PyGithub
slackclient
python-dotenv
pandas
anthropic
slack_sdk
splitio_client
setuptools